# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "This Api for Photocontest App",
        "description": "Provide all the service for the Photocontest app",
        "termsOfService": "",
        "contact": {
            "email": "abdalrahman.alsharawy@gmail.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost",
    "basePath": "/",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
        "/items/approve": {
            "post": {
                "tags": [
                    "items"
                ],
                "description": "Approve an item.",
                "operationId": "approve",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Token",
                        "in": "header",
                        "description": "Admin token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "item_id",
                        "in": "formData",
                        "description": "",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "items List."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
        "/items/disapprove": {
            "post": {
                "tags": [
                    "items"
                ],
                "description": "disapprove an item.",
                "operationId": "disapprove",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Token",
                        "in": "header",
                        "description": "Admin token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "item_id",
                        "in": "formData",
                        "description": "",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "item."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
        "/user/items": {
            "get": {
                "tags": [
                    "users"
                ],
                "description": "Returns List of user's items.",
                "operationId": "user_items",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Token",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "items List."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
        "/items": {
            "get": {
                "tags": [
                    "items"
                ],
                "description": "Returns List of items.",
                "operationId": "index",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "items List."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            },
            "post": {
                "tags": [
                    "items"
                ],
                "description": "create item.",
                "operationId": "store",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Token",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "price",
                        "in": "formData",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "available_stock",
                        "in": "formData",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "image_url",
                        "in": "formData",
                        "description": "Base64 String",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "category_id",
                        "in": "formData",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "created."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
        "/votes": {
            "get": {
                "tags": [
                    "votes"
                ],
                "description": "Returns List of votes.",
                "operationId": "votes",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "votes List."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
        "/admin/items": {
            "get": {
                "tags": [
                    "admin"
                ],
                "description": "Returns List of items for admin pannel.",
                "operationId": "index",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Token",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "items List."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
        "/items/{item}": {
            "get": {
                "tags": [
                    "items"
                ],
                "description": "Returns item.",
                "operationId": "show",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "item",
                        "in": "path",
                        "description": "ID of item",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "item object."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            },
            "put": {
                "tags": [
                    "items"
                ],
                "description": "create item.",
                "operationId": "update",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Token",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "item",
                        "in": "path",
                        "description": "ID of item",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "formData",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "price",
                        "in": "formData",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "available_stock",
                        "in": "formData",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "image_url",
                        "in": "formData",
                        "description": "Base64 string",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "created."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            },
            "delete": {
                "tags": [
                    "items"
                ],
                "description": "Delete item.",
                "operationId": "destroy",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Token",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "item",
                        "in": "path",
                        "description": "ID of item",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "item object."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
        "/items/vote": {
            "post": {
                "tags": [
                    "items"
                ],
                "description": "Returns total likes.",
                "operationId": "like",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "item_id",
                        "in": "formData",
                        "description": "ID of item",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "value",
                        "in": "formData",
                        "description": "user's vote",
                        "required": true,
                        "type": "boolean"
                    },
                    {
                        "name": "Token",
                        "in": "header",
                        "description": "user's token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "total likes."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
...
    },
...
}